css/css-scroll-snap/parsing/scroll-margin-top-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-margin-top with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-margin-top">
<meta name="assert" content="scroll-margin-top supports only the grammar '<length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("scroll-margin-top", "auto");

test_invalid_value("scroll-margin-top", "20%");
test_invalid_value("scroll-margin-top", "1px 2px");
</script>
</body>
</html>
